рефакторинг. начало
